requirements.txt
# Core AI/LLM
anthropic==0.18.1
langchain==0.1.0
langchain-anthropic==0.1.1
langchain-community==0.0.13
langchain-core==0.1.10

# Document Processing
pypdf==3.17.4
pdfplumber==0.10.3
unstructured==0.11.6
python-docx==1.1.0
python-magic-bin==0.4.14  # Use python-magic==0.4.27 for Linux/Mac

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Web Scraping & Search
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
duckduckgo-search==4.1.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0